chore: add x5c verification for 1.3 SD-JWT
